refactor: update logger import paths to use app module
