+++ .idea/dictionaries/project.xml
@@ -38,6 +38,7 @@
       <w>hstack</w>
       <w>jlumbroso</w>
       <w>latentdiffusion</w>
+      <w>librosa</w>
       <w>logvar</w>
       <w>lrelu</w>
       <w>markdownlint</w>
