+++ .idea/misc.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="Black">
-    <option name="sdkName" value="Python 3.10 (VideoGenerator)" />
+    <option name="sdkName" value="uv (visualizr)" />
   </component>
   <component name="ProjectRootManager" version="2" project-jdk-name="uv (visualizr)" project-jdk-type="Python SDK" />
 </project>
