+++ .idea/ruff.xml
@@ -3,8 +3,8 @@
   <component name="RuffConfigService">
     <option name="alwaysUseGlobalRuff" value="true" />
     <option name="enableRuffLogging" value="true" />
-    <option name="globalRuffExecutablePath" value="C:\Users\moham\AppData\Local\Microsoft\WinGet\Packages\astral-sh.ruff_Microsoft.Winget.Source_8wekyb3d8bbwe\ruff.exe" />
-    <option name="ruffConfigPath" value="D:\Sync\Development\visualizr\.trunk\configs\ruff.toml" />
+    <option name="globalRuffExecutablePath" value="C:\Users\Mohamed\AppData\Local\Microsoft\WinGet\Packages\astral-sh.ruff_Microsoft.Winget.Source_8wekyb3d8bbwe\ruff.exe" />
+    <option name="ruffConfigPath" value="D:\Development\visualizr\.trunk\configs\ruff.toml" />
     <option name="runRuffOnSave" value="true" />
     <option name="useIntellijLspClient" value="false" />
     <option name="useLsp4ij" value="true" />
