+++ src/visualizr/anitalker/networks/encoder.py
@@ -4,7 +4,7 @@ from gradio import Info
 from torch import Tensor, flip, float32, nn, randn, softmax, tensor, zeros, zeros_like
 from torch.nn.functional import conv2d, leaky_relu, linear, pad
 
-from visualizr import logger
+from visualizr.app.logger import logger
 
 
 def fused_leaky_relu(_input, bias, negative_slope=0.2, scale=2**0.5):
