+++ src/visualizr/app/runner.py
@@ -1,5 +1,5 @@
 from visualizr.app.builder import App
-from visualizr.settings import Settings
+from visualizr.app.settings import Settings
 
 settings: Settings = Settings()
 app: App = App(settings)
